[FAB2023-989] fix: 코드 리뷰에 따른 2차 코드 정리
